database-management/postgres/requirements.txt
psycopg2-binary
meilisearch
